implement detect-faces express route
